[DEV-704] slight renaming of integration factories
