--- src/subdomains/core/liquidity-management/factories/liquidity-balance.factory.ts
@@ -1,10 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { LiquidityManagementRule } from '../entities/liquidity-management-rule.entity';
-import { LiquidityBalanceProvider } from '../interfaces';
-
-@Injectable()
-export class LiquidityBalanceFactory {
-  getIntegration(rule: LiquidityManagementRule): LiquidityBalanceProvider {
-    return null;
-  }
-}
