+++ src/subdomains/core/liquidity-management/interfaces/index.ts
@@ -2,7 +2,7 @@ import { Asset } from 'src/shared/models/asset/asset.entity';
 import { Fiat } from 'src/shared/models/fiat/fiat.entity';
 import { LiquidityBalance } from '../entities/liquidity-balance.entity';
 
-export interface LiquidityBalanceProvider {
+export interface LiquidityBalanceIntegration {
   getBalance(asset: Asset | Fiat): Promise<LiquidityBalance>;
 }
