+++ src/subdomains/core/liquidity-management/services/liquidity-management-rule.service.ts
@@ -12,8 +12,7 @@ import { LiquidityManagementActionRepository } from '../repositories/liquidity-m
 import { LiquidityManagementRuleOutputDto } from '../dto/output/liquidity-management-rule-output.dto';
 import { LiquidityManagementRuleOutputDtoMapper } from '../dto/output/mappers/liquidity-management-rule-output-dto.mapper';
 import { LiquidityManagementRuleCreationDto } from '../dto/input/liquidity-management-rule-creation.dto';
-import { LiquidityActionIntegration } from '../interfaces';
-import { LiquidityActionIntegrationFactory } from '../factories/liquidity-action.integration.factory';
+import { LiquidityActionIntegrationFactory } from '../factories/liquidity-action-integration.factory';
 
 @Injectable()
 export class LiquidityManagementRuleService {
@@ -22,7 +21,7 @@ export class LiquidityManagementRuleService {
     private readonly actionRepo: LiquidityManagementActionRepository,
     private readonly assetService: AssetService,
     private readonly fiatService: FiatService,
-    private readonly liquidityActionIntegrationFactory: LiquidityActionIntegrationFactory,
+    private readonly actionIntegrationFactory: LiquidityActionIntegrationFactory,
   ) {}
 
   //*** PUBLIC API ***//
@@ -74,10 +73,6 @@ export class LiquidityManagementRuleService {
     rule.reactivate();
 
     return LiquidityManagementRuleOutputDtoMapper.entityToDto(await this.ruleRepo.save(rule));
-  }
-
-  findLiquidityActionIntegration(action: LiquidityManagementAction): LiquidityActionIntegration {
-    return this.liquidityActionIntegrationFactory.getIntegration(action);
   }
 
   //*** HELPER METHODS ***//
@@ -185,7 +180,7 @@ export class LiquidityManagementRuleService {
       actionOnFail,
     );
 
-    const integration = this.findLiquidityActionIntegration(newAction);
+    const integration = this.actionIntegrationFactory.getIntegration(newAction);
 
     if (!integration) {
       throw new BadRequestException(
